fix: 적립률 표시 방식 수정

+++ src/main/java/com/tn/member/service/MemberService.java
@@ -72,7 +72,7 @@ public interface MemberService {
 
 	boolean removeAddress(int addressId) throws Exception;
 
-	double getPointRate(String userId) throws Exception;
+	int getPointRate(String userId) throws Exception;
