+++ src/main/java/com/tn/member/service/MemberServiceImpl.java
@@ -277,9 +277,9 @@ public class MemberServiceImpl implements MemberService {
 
 
 	@Override
-	public double getPointRate(String userId) throws Exception {
+	public int getPointRate(String userId) throws Exception {
 		
-		double rate = 0;
+		int rate = 0;
 		
 		if (userId != null) {
 			rate = dao.getPointRate(userId);
